Improved Rides list view, now showing all needed data Including  Drivers' Name and their reputation

+++ app/src/main/java/com/example/carpooldummyapp/PassengerAccount.java
@@ -5,7 +5,7 @@ public class PassengerAccount extends Account {
     private ACCOUNT_TYPE Account_Type = ACCOUNT_TYPE.PASSENGER;
 
 
-    public PassengerAccount(String name, String lastname, String phone, String Email, int rating, ACCOUNT_TYPE account_type) {
+    public PassengerAccount(String name, String lastname, String phone, String Email, float rating, ACCOUNT_TYPE account_type) {
         super(name, lastname, phone, Email, rating, account_type);
     }
